# [Enhancement] improve struct performance