--- a/be/src/storage/rowset/struct_column_iterator.cpp
+++ b/be/src/storage/rowset/struct_column_iterator.cpp
@@ -14,10 +14,13 @@
 
 #include "storage/rowset/struct_column_iterator.h"
 
+#include <unordered_map>
+
 #include "column/column_access_path.h"
 #include "column/const_column.h"
 #include "column/nullable_column.h"
 #include "column/struct_column.h"
+#include "storage/rowset/column_iterator.h"
 #include "storage/rowset/column_reader.h"
 #include "storage/rowset/scalar_column_iterator.h"
 
@@ -64,7 +67,9 @@ private:
     std::vector<std::unique_ptr<ColumnIterator>> _field_iters;
     const ColumnAccessPath* _path;
 
-    std::vector<uint8_t> _access_flags;
+    // prune subfield by path
+    std::vector<ColumnIterator*> _access_iters;
+    std::unordered_map<int, int> _access_index_map;
 };
 
 StatusOr<std::unique_ptr<ColumnIterator>> create_struct_iter(ColumnReader* _reader,
@@ -90,12 +95,22 @@ Status StructColumnIterator::init(const ColumnIteratorOptions& opts) {
     }
 
     if (_path != nullptr && !_path->children().empty()) {
-        _access_flags.resize(_field_iters.size(), 0);
+        std::vector<ColumnAccessPath*> child_paths(_field_iters.size(), nullptr);
         for (const auto& child : _path->children()) {
-            _access_flags[child->index()] = 1;
+            child_paths[child->index()] = child.get();
+        }
+
+        for (int i = 0; i < _field_iters.size(); ++i) {
+            if (child_paths[i] != nullptr) {
+                _access_index_map[i] = _access_iters.size();
+                _access_iters.push_back(_field_iters[i].get());
+            }
         }
     } else {
-        _access_flags.resize(_field_iters.size(), 1);
+        for (int i = 0; i < _field_iters.size(); ++i) {
+            _access_index_map[i] = _access_iters.size();
+            _access_iters.push_back(_field_iters[i].get());
+        }
     }
 
     return Status::OK();
@@ -119,24 +134,11 @@ Status StructColumnIterator::next_batch(size_t* n, Column* dst) {
         down_cast<NullableColumn*>(dst)->update_has_null();
     }
 
-    size_t row_count = 0;
+    DCHECK_EQ(struct_column->fields_column().size(), _access_iters.size());
     auto& fields = struct_column->fields_column();
-    for (int i = 0; i < _field_iters.size(); ++i) {
-        if (_access_flags[i]) {
-            auto num_to_read = *n;
-            RETURN_IF_ERROR(_field_iters[i]->next_batch(&num_to_read, fields[i].get()));
-            row_count = fields[i]->size();
-        }
-    }
-
-    for (int i = 0; i < _field_iters.size(); ++i) {
-        if (!_access_flags[i]) {
-            if (!fields[i]->is_constant()) {
-                fields[i]->append_default(1);
-                fields[i] = ConstColumn::create(fields[i], 1);
-            }
-            fields[i]->resize(row_count);
-        }
+    for (int i = 0; i < _access_iters.size(); ++i) {
+        auto num_to_read = *n;
+        RETURN_IF_ERROR(_access_iters[i]->next_batch(&num_to_read, fields[i].get()));
     }
 
     return Status::OK();
@@ -158,25 +160,14 @@ Status StructColumnIterator::next_batch(const SparseRange<>& range, Column* dst)
         RETURN_IF_ERROR(_null_iter->next_batch(range, null_column));
         down_cast<NullableColumn*>(dst)->update_has_null();
     }
+
+    DCHECK_EQ(struct_column->fields_column().size(), _access_iters.size());
     // Read all fields
-    size_t row_count = 0;
     auto& fields = struct_column->fields_column();
-    for (int i = 0; i < _field_iters.size(); ++i) {
-        if (_access_flags[i]) {
-            RETURN_IF_ERROR(_field_iters[i]->next_batch(range, fields[i].get()));
-            row_count = fields[i]->size();
-        }
+    for (int i = 0; i < _access_iters.size(); ++i) {
+        RETURN_IF_ERROR(_access_iters[i]->next_batch(range, fields[i].get()));
     }
 
-    for (int i = 0; i < _field_iters.size(); ++i) {
-        if (!_access_flags[i]) {
-            if (!fields[i]->is_constant()) {
-                fields[i]->append_default(1);
-                fields[i] = ConstColumn::create(fields[i], 1);
-            }
-            fields[i]->resize(row_count);
-        }
-    }
     return Status::OK();
 }
 
@@ -194,25 +185,13 @@ Status StructColumnIterator::fetch_values_by_rowid(const rowid_t* rowids, size_t
         struct_column = down_cast<StructColumn*>(values);
     }
 
+    DCHECK_EQ(struct_column->fields_column().size(), _access_iters.size());
     // read all fields
     auto& fields = struct_column->fields_column();
-    size_t row_count = 0;
-    for (int i = 0; i < _field_iters.size(); ++i) {
-        if (_access_flags[i]) {
-            RETURN_IF_ERROR(_field_iters[i]->fetch_values_by_rowid(rowids, size, fields[i].get()));
-            row_count = fields[i]->size();
-        }
+    for (int i = 0; i < _access_iters.size(); ++i) {
+        RETURN_IF_ERROR(_access_iters[i]->fetch_values_by_rowid(rowids, size, fields[i].get()));
     }
 
-    for (int i = 0; i < _field_iters.size(); ++i) {
-        if (!_access_flags[i]) {
-            if (!fields[i]->is_constant()) {
-                fields[i]->append_default(1);
-                fields[i] = ConstColumn::create(fields[i], 1);
-            }
-            fields[i]->resize(row_count);
-        }
-    }
     return Status::OK();
 }
 
@@ -220,7 +199,7 @@ Status StructColumnIterator::seek_to_first() {
     if (_null_iter != nullptr) {
         RETURN_IF_ERROR(_null_iter->seek_to_first());
     }
-    for (auto& iter : _field_iters) {
+    for (auto& iter : _access_iters) {
         RETURN_IF_ERROR(iter->seek_to_first());
     }
     return Status::OK();
@@ -230,7 +209,7 @@ Status StructColumnIterator::seek_to_ordinal(ordinal_t ord) {
     if (_null_iter != nullptr) {
         RETURN_IF_ERROR(_null_iter->seek_to_ordinal(ord));
     }
-    for (auto& iter : _field_iters) {
+    for (auto& iter : _access_iters) {
         RETURN_IF_ERROR(iter->seek_to_ordinal(ord));
     }
     return Status::OK();
@@ -251,11 +230,12 @@ Status StructColumnIterator::next_batch(size_t* n, Column* dst, ColumnAccessPath
     // 1. init predicate access path
     std::vector<uint8_t> predicate_access_flags;
     std::vector<ColumnAccessPath*> predicate_child_paths;
-    predicate_access_flags.resize(_field_iters.size(), 0);
-    predicate_child_paths.resize(_field_iters.size(), nullptr);
+    predicate_access_flags.resize(_access_iters.size(), 0);
+    predicate_child_paths.resize(_access_iters.size(), nullptr);
     for (const auto& child : path->children()) {
-        predicate_access_flags[child->index()] = 1;
-        predicate_child_paths[child->index()] = child.get();
+        auto index = _access_index_map[child->index()];
+        predicate_access_flags[index] = 1;
+        predicate_child_paths[index] = child.get();
     }
 
     StructColumn* struct_column = nullptr;
@@ -274,23 +254,21 @@ Status StructColumnIterator::next_batch(size_t* n, Column* dst, ColumnAccessPath
         down_cast<NullableColumn*>(dst)->update_has_null();
     }
 
+    DCHECK_EQ(struct_column->fields_column().size(), _access_iters.size());
     // 3. Read fields
     size_t row_count = 0;
     auto& fields = struct_column->fields_column();
-    for (int i = 0; i < _field_iters.size(); ++i) {
+    for (int i = 0; i < _access_iters.size(); ++i) {
         if (predicate_access_flags[i]) {
             auto num_to_read = *n;
-            RETURN_IF_ERROR(_field_iters[i]->next_batch(&num_to_read, fields[i].get(), predicate_child_paths[i]));
+            RETURN_IF_ERROR(_access_iters[i]->next_batch(&num_to_read, fields[i].get(), predicate_child_paths[i]));
             row_count = fields[i]->size();
         }
     }
 
-    for (int i = 0; i < _field_iters.size(); ++i) {
+    for (int i = 0; i < _access_iters.size(); ++i) {
         if (!predicate_access_flags[i]) {
-            if (!fields[i]->is_constant()) {
-                fields[i]->append_default(1);
-                fields[i] = ConstColumn::create(fields[i], 1);
-            }
+            DCHECK(fields[i]->is_constant());
             fields[i]->resize(row_count);
         }
     }
@@ -304,11 +282,13 @@ Status StructColumnIterator::next_batch(const SparseRange<>& range, Column* dst,
 
     std::vector<uint8_t> predicate_access_flags;
     std::vector<ColumnAccessPath*> predicate_child_paths;
-    predicate_access_flags.resize(_field_iters.size(), 0);
-    predicate_child_paths.resize(_field_iters.size(), nullptr);
+
+    predicate_access_flags.resize(_access_iters.size(), 0);
+    predicate_child_paths.resize(_access_iters.size(), nullptr);
     for (const auto& child : path->children()) {
-        predicate_access_flags[child->index()] = 1;
-        predicate_child_paths[child->index()] = child.get();
+        auto index = _access_index_map[child->index()];
+        predicate_access_flags[index] = 1;
+        predicate_child_paths[index] = child.get();
     }
 
     StructColumn* struct_column = nullptr;
@@ -329,15 +309,15 @@ Status StructColumnIterator::next_batch(const SparseRange<>& range, Column* dst,
     // Read all fields
     size_t row_count = 0;
     auto& fields = struct_column->fields_column();
-    for (int i = 0; i < _field_iters.size(); ++i) {
+    for (int i = 0; i < _access_iters.size(); ++i) {
         if (!predicate_access_flags[i]) {
             continue;
         }
-        RETURN_IF_ERROR(_field_iters[i]->next_batch(range, fields[i].get(), predicate_child_paths[i]));
+        RETURN_IF_ERROR(_access_iters[i]->next_batch(range, fields[i].get(), predicate_child_paths[i]));
         row_count = fields[i]->size();
     }
 
-    for (int i = 0; i < _field_iters.size(); ++i) {
+    for (int i = 0; i < _access_iters.size(); ++i) {
         if (!predicate_access_flags[i]) {
             if (!fields[i]->is_constant()) {
                 fields[i]->append_default(1);
@@ -359,32 +339,20 @@ Status StructColumnIterator::fetch_subfield_by_rowid(const rowid_t* rowids, size
         struct_column = down_cast<StructColumn*>(values);
     }
 
+    DCHECK_EQ(struct_column->fields_column().size(), _access_iters.size());
     // read all fields
     auto& fields = struct_column->fields_column();
-    size_t row_count = 0;
-    for (int i = 0; i < _field_iters.size(); ++i) {
-        if (_access_flags[i]) {
-            if (fields[i]->is_constant()) {
-                // doesn't meterialized
-                fields[i] = down_cast<ConstColumn*>(fields[i].get())->data_column()->clone_empty();
-                RETURN_IF_ERROR(_field_iters[i]->fetch_values_by_rowid(rowids, size, fields[i].get()));
-            } else {
-                // handle nested struct
-                // structA -> structB -> e (meterialized)
-                //                    -> f (un-meterialized)
-                RETURN_IF_ERROR(_field_iters[i]->fetch_subfield_by_rowid(rowids, size, fields[i].get()));
-            }
-            row_count = fields[i]->size();
-        }
-    }
-
-    for (int i = 0; i < _field_iters.size(); ++i) {
-        if (!_access_flags[i]) {
-            if (!fields[i]->is_constant()) {
-                fields[i]->append_default(1);
-                fields[i] = ConstColumn::create(fields[i], 1);
-            }
-            fields[i]->resize(row_count);
+    for (int i = 0; i < _access_iters.size(); ++i) {
+        if (fields[i]->is_constant()) {
+            // doesn't meterialized
+            fields[i] = down_cast<ConstColumn*>(fields[i].get())->data_column();
+            fields[i]->resize_uninitialized(0);
+            RETURN_IF_ERROR(_access_iters[i]->fetch_values_by_rowid(rowids, size, fields[i].get()));
+        } else {
+            // handle nested struct
+            // structA -> structB -> e (meterialized)
+            //                    -> f (un-meterialized)
+            RETURN_IF_ERROR(_access_iters[i]->fetch_subfield_by_rowid(rowids, size, fields[i].get()));
         }
     }
     return Status::OK();