--- a/be/test/storage/rowset/struct_column_rw_test.cpp
+++ b/be/test/storage/rowset/struct_column_rw_test.cpp
@@ -196,7 +196,7 @@ protected:
                 ASSERT_TRUE(st.ok());
                 ASSERT_EQ(src_column->size(), rows_read);
 
-                ASSERT_EQ("{f1:1,f2:CONST: ''}", dst_column->debug_item(0));
+                ASSERT_EQ("{f1:1,f2:''}", dst_column->debug_item(0));
             }
         }