## Why I'm doing:

## What I'm doing:
Improve struct performance, to prune unused subfield in schema, then the `xx_iterator` dones't init unused subfield
save 20%~50% costs

test case: 
SSB 100GB

|   WITH_PREDICATE  |   ORIGN   |   ORIGIN(DISABLE low_cardinality) |   STRUCT(BEFORE PR)   |   STRUCT(THIS PR) |   STRUCT BEFORE/PR   |   ORIGIN/PR  |
|   ----    |   ----    |   ----    |   ----    |   ----    |   ----    |   ----    |
|   Q01 |   84  |   82  |   404 |   141 |   2.87    |   1.68    |
|   Q02 |   16  |   16  |   57  |   24  |   2.38    |   1.50    |
|   Q03 |   58  |   83  |   151 |   113 |   1.34    |   1.95    |
|   Q04 |   830 |   840 |   3821    |   2733    |   1.40    |   3.29    |
|   Q05 |   686 |   678 |   3421    |   2205    |   1.55    |   3.21    |
|   Q06 |   525 |   517 |   2747    |   1912    |   1.44    |   3.64    |
|   Q07 |   779 |   870 |   3695    |   2642    |   1.40    |   3.39    |
|   Q08 |   631 |   652 |   2950    |   1848    |   1.60    |   2.93    |
|   Q09 |   432 |   428 |   1986    |   1585    |   1.25    |   3.67    |
|   Q10 |   21  |   20  |   58  |   40  |   1.45    |   1.90    |
|   Q11 |   975 |   1009    |   5077    |   4019    |   1.26    |   4.12    |
|   Q12 |   314 |   337 |   1427    |   1133    |   1.26    |   3.61    |
|   Q13 |   242 |   248 |   886 |   619 |   1.43    |   2.56    |

Q4~Q9 looks struct query is more poor than scalar, I think the reason is that the struct predicate can't use zone_map

then I remove all struct predicate and scalar predicate, and disable low_cardinality:

|   NO_PREDICATE    |   ORIGN   |   ORIGIN(DISABLE low_cardinality) |   STRUCT(BEFORE PR)   |   STRUCT(THIS PR) |   STRUCT BEFORE/PR    |   ORIGIN/PR  |
|   ----    |   ----    |   ----    |   ----    |   ----    |   ----    |   ----    |
|   Q01 |   83  |   80  |   350 |   113 |   3.10    |   1.41    |
|   Q02 |   19  |   14  |   53  |   20  |   2.65    |   1.43    |
|   Q03 |   63  |   60  |   87  |   64  |   1.36    |   1.07    |
|   Q04 |   2097    |   2146    |   3721    |   2635    |   1.41    |   1.23    |
|   Q05 |   2123    |   2087    |   3692    |   2666    |   1.38    |   1.28    |
|   Q06 |   2120    |   2108    |   3632    |   2679    |   1.36    |   1.27    |
|   Q07 |   1308    |   3850    |   4215    |   3378    |   1.25    |   0.88    |
|   Q08 |   6560    |   12596   |   14337   |   12842   |   1.12    |   1.02    |
|   Q09 |   6396    |   12485   |   14247   |   12909   |   1.10    |   1.03    |

it's looks more similar in Q8~Q9, but Q4 ~ Q7 diff 600ms. Through the flame graph, I think the reason is that the struct contains null values, while the scalar column doesn't

|   NO_PREDICATE    |   ORIGN(Nullable) |   ORIGIN(DISABLE low_cardinality) |   STRUCT(BEFORE PR)   |   STRUCT(THIS PR) |   BEFORE/PR    |   ORIGIN/PR    |
|   ----    |   ----    |   ----    |   ----    |   ----    |   ----    |   ----    |
|   Q01 |   91  |   94  |   350 |   113 |   3.10    |   1.20    |
|   Q02 |   16  |   18  |   53  |   20  |   2.65    |   1.11    |
|   Q03 |   63  |   65  |   87  |   64  |   1.36    |   0.98    |
|   Q04 |   2549    |   2589    |   3721    |   2635    |   1.41    |   1.02    |
|   Q05 |   2583    |   2617    |   3692    |   2666    |   1.38    |   1.02    |
|   Q06 |   2591    |   2581    |   3632    |   2679    |   1.36    |   1.04    |
|   Q07 |   1645    |   4108    |   4215    |   3378    |   1.25    |   0.82    |
|   Q08 |   7132    |   13436   |   14337   |   12842   |   1.12    |   0.96    |
|   Q09 |   7101    |   13841   |   14247   |   12909   |   1.10    |   0.93    |

now the performance looks similar.

next performance optimization point:
* zone_map_filter on struct
* low_cardinality on struct
* nullable on struct

Fixes #issue

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5